creates two flavors
