+++ common/build.gradle
@@ -7,7 +7,8 @@ dependencies {
     implementation project(":strings")
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlinVersion"
     implementation 'androidx.core:core-ktx:1.3.1'
-    implementation "com.android.billingclient:billing:$billingVersion"
+    newBillingImplementation "com.android.billingclient:billing:$billingVersion"
+    oldBillingImplementation "com.android.billingclient:billing:3.0.2"
     testImplementation project(":test-utils")
     testImplementation 'androidx.test:core:1.2.0'
     testImplementation 'androidx.test:runner:1.2.0'
@@ -16,7 +17,8 @@ dependencies {
     testImplementation 'org.robolectric:robolectric:4.3'
     testImplementation 'com.squareup.okhttp3:mockwebserver:4.2.0'
     testImplementation 'org.mockito:mockito-core:3.0.0'
-    testImplementation "com.android.billingclient:billing:$billingVersion"
+    testNewBillingImplementation "com.android.billingclient:billing:$billingVersion"
+    testOldBillingImplementation "com.android.billingclient:billing:3.0.2"
     testImplementation 'io.mockk:mockk:1.10.0'
     testImplementation 'org.assertj:assertj-core:3.13.2'
 }
