+++ examples/purchase-tester-java/build.gradle
@@ -9,7 +9,15 @@ android {
 
         vectorDrawables.useSupportLibrary = true
     }
-
+    flavorDimensions "billing"
+    productFlavors {
+        oldBilling {
+            dimension "billing"
+        }
+        newBilling {
+            dimension "billing"
+        }
+    }
     buildTypes {
         release {
             minifyEnabled true
