+++ feature/google/build.gradle
@@ -11,7 +11,8 @@ dependencies {
     implementation 'com.google.android.gms:play-services-ads-identifier:17.0.1'
     implementation "androidx.core:core-ktx:1.3.1"
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlinVersion"
-    implementation "com.android.billingclient:billing:$billingVersion"
+    newBillingImplementation "com.android.billingclient:billing:$billingVersion"
+    oldBillingImplementation "com.android.billingclient:billing:3.0.2"
 
     testImplementation project(":test-utils")
     testImplementation 'com.google.android.gms:play-services-ads-identifier:17.0.1'
