+++ feature/google/src/oldBilling/java/com/revenuecat/purchases/google/BillingFlowParamsExtensions.kt
@@ -1,0 +1,11 @@
+package com.revenuecat.purchases.google
+
+import com.android.billingclient.api.BillingFlowParams
+import com.revenuecat.purchases.common.ReplaceSkuInfo
+
+fun BillingFlowParams.Builder.setUpgradeInfo(replaceSkuInfo: ReplaceSkuInfo) {
+    setOldSku(replaceSkuInfo.oldPurchase.sku, replaceSkuInfo.oldPurchase.purchaseToken)
+    replaceSkuInfo.prorationMode?.let { prorationMode ->
+        setReplaceSkusProrationMode(prorationMode)
+    }
+}
