+++ integration-tests/build.gradle
@@ -17,6 +17,15 @@ android {
             keyPassword project.properties['releaseKeyPassword']
         }
     }
+    flavorDimensions "billing"
+    productFlavors {
+        oldBilling {
+            dimension "billing"
+        }
+        newBilling {
+            dimension "billing"
+        }
+    }
     buildTypes {
         release {
             minifyEnabled true
