+++ library.gradle
@@ -5,7 +5,15 @@ apply plugin: 'kotlin-android-extensions'
 android {
     compileSdkVersion compileVersion
     buildToolsVersion buildToolsVersion
-
+    flavorDimensions "billing"
+    productFlavors {
+        oldBilling {
+            dimension "billing"
+        }
+        newBilling {
+            dimension "billing"
+        }
+    }
     defaultConfig {
         minSdkVersion minVersion
         targetSdkVersion compileVersion
@@ -43,9 +51,14 @@ dependencies {
     testImplementation 'org.robolectric:robolectric:4.6.1'
     testImplementation 'com.squareup.okhttp3:mockwebserver:4.2.0'
     testImplementation 'org.mockito:mockito-core:3.0.0'
-    testImplementation "com.android.billingclient:billing:$billingVersion"
+    testNewBillingImplementation "com.android.billingclient:billing:$billingVersion"
+    testOldBillingImplementation "com.android.billingclient:billing:3.0.2"
     testImplementation 'io.mockk:mockk:1.10.0'
     testImplementation 'org.assertj:assertj-core:3.13.2'
 }
 
 apply plugin: 'com.vanniktech.maven.publish'
+
+mavenPublish {
+    androidVariantToPublish = "newBillingRelease"
+}
