+++ purchases/build.gradle
@@ -14,7 +14,8 @@ dependencies {
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlinVersion"
     implementation 'androidx.annotation:annotation:1.1.0'
-    api "com.android.billingclient:billing:$billingVersion"
+    newBillingApi "com.android.billingclient:billing:$billingVersion"
+    oldBillingApi "com.android.billingclient:billing:3.0.2"
     implementation "androidx.lifecycle:lifecycle-runtime:$lifecycleVersion"
     kapt "androidx.lifecycle:lifecycle-compiler:$lifecycleVersion"
     implementation "androidx.lifecycle:lifecycle-process:$lifecycleVersion"
@@ -27,7 +28,8 @@ dependencies {
     testImplementation 'org.robolectric:robolectric:4.3'
     testImplementation 'com.squareup.okhttp3:mockwebserver:4.2.0'
     testImplementation 'org.mockito:mockito-core:3.0.0'
-    testImplementation "com.android.billingclient:billing:$billingVersion"
+    testNewBillingImplementation "com.android.billingclient:billing:$billingVersion"
+    testOldBillingImplementation "com.android.billingclient:billing:3.0.2"
     testImplementation 'io.mockk:mockk:1.10.0'
     testImplementation 'org.assertj:assertj-core:3.13.2'
 }
