+++ test-utils/build.gradle
@@ -4,7 +4,8 @@ dependencies {
     implementation fileTree(dir: "libs", include: ["*.jar"])
     implementation project(":common")
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlinVersion"
-    implementation "com.android.billingclient:billing:$billingVersion"
+    newBillingImplementation "com.android.billingclient:billing:$billingVersion"
+    oldBillingImplementation "com.android.billingclient:billing:3.0.2"
     implementation 'org.assertj:assertj-core:3.13.2'
     implementation 'io.mockk:mockk:1.10.0'
     implementation 'androidx.test.ext:junit:1.1.2'
